created generalized "Submitter" for all strategies
